Used safer injection approach in for Specflow components
